Drop support for elasticsearch 7 and elasticsearch 8.14 - closes #830
