--- .github/actions/elastic_7/action.yml
@@ -1,42 +1,0 @@
-name: 'Pytest run'
-description: 'Greet someone'
-inputs:
-  python-version:  # id of input
-    description: 'Python version to use'
-    required: true
-  elasticsearch:
-    description: 'Elasticsearch version'
-    required: true
-runs:
-  using: "composite"
-  steps:
-  - name: Set up Python ${{ inputs.python-version }}
-    uses: actions/setup-python@v5
-    with:
-      python-version: ${{ inputs.python-version }}
-  - uses: actions/cache@v4
-    with:
-      path: ~/elasticsearch
-      key: elasticsearch-${{ inputs.elasticsearch }}
-  - name: Setup elasticsearch
-    uses: ankane/setup-elasticsearch@v1
-    with:
-      elasticsearch-version: ${{ inputs.elasticsearch }}
-  - name: Prep environment
-    shell: bash
-    run: pip install pipenv && pipenv install --dev
-  - name: Downgrade elasticsearch
-    shell: bash
-    run: pipenv run pip install "elasticsearch<8"
-  - name: Run test
-    shell: bash
-    run: pipenv run pytest --elasticsearch-executable=$ES_HOME/bin/elasticsearch -n 0 --cov-report=xml
-  - name: Run xdist test
-    shell: bash
-    run: pipenv run pytest --elasticsearch-executable=$ES_HOME/bin/elasticsearch -n 1 --cov-report=xml:coverage-xdist.xml
-  - name: Upload coverage to Codecov
-    uses: codecov/codecov-action@v5
-    with:
-      flags: linux,elastic-${{ inputs.elasticsearch }}
-      env_vars: OS, PYTHON
-      fail_ci_if_error: false
