+++ .github/actions/pytest/action.yml
@@ -1,5 +1,5 @@
 name: 'Pytest run'
-description: 'Greet someone'
+description: 'Runs recent elasticsearch version'
 inputs:
   python-version:  # id of input
     description: 'Python version to use'
@@ -11,10 +11,10 @@ runs:
   using: "composite"
   steps:
   - name: Set up Python ${{ inputs.python-version }}
-    uses: actions/setup-python@v5
+    uses: actions/setup-python@v6
     with:
       python-version: ${{ inputs.python-version }}
-  - uses: actions/cache@v4
+  - uses: actions/cache@v5
     with:
       path: ~/elasticsearch
       key: elasticsearch-${{ inputs.elasticsearch }}
@@ -23,15 +23,17 @@ runs:
     with:
       elasticsearch-version: ${{ inputs.elasticsearch }}
   - name: Run test
-    uses: fizyk/actions-reuse/.github/actions/pipenv@v2.4.8
+    uses: fizyk/actions-reuse/.github/actions/pipenv@v4.1.1
     with:
-      python-version: ${{ matrix.python-version }}
+      python-version: ${{ inputs.python-version }}
       command: pytest --elasticsearch-executable=$ES_HOME/bin/elasticsearch -n 0 --cov-report=xml
+      allow-prereleases: true
   - name: Run xdist test
-    uses: fizyk/actions-reuse/.github/actions/pipenv@v2.4.8
+    uses: fizyk/actions-reuse/.github/actions/pipenv@v4.1.1
     with:
-      python-version: ${{ matrix.python-version }}
+      python-version: ${{ inputs.python-version }}
       command: pytest --elasticsearch-executable=$ES_HOME/bin/elasticsearch -n 1 --cov-report=xml:coverage-xdist.xml
+      allow-prereleases: true
   - name: Upload coverage to Codecov
     uses: codecov/codecov-action@v5
     with:
